requirements.txt
Flask==3.1.2
Flask-Cors==6.0.2
Flask-SQLAlchemy==3.1.1
mysql-connector-python==9.5.0
PyMySQL==1.1.2
SQLAlchemy==2.0.45
Werkzeug==3.1.5
cryptography==42.0.8
python-dotenv
openpyxl